refactor: Extract transform code to @jest-transform
