+++ packages/jest-core/src/runGlobalHook.ts
@@ -5,9 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {extname} from 'path';
 import pEachSeries from 'p-each-series';
-import {addHook} from 'pirates';
 import {Config} from '@jest/types';
 import {Test} from 'jest-runner';
 import {ScriptTransformer} from '@jest/transform';
@@ -47,36 +45,9 @@ export default async ({
 
       const transformer = new ScriptTransformer(projectConfig);
 
-      // Load the transformer to avoid a cycle where we need to load a
-      // transformer in order to transform it in the require hooks
-      transformer.preloadTransformer(modulePath);
-
-      let transforming = false;
-      const revertHook = addHook(
-        (code, filename) => {
-          try {
-            transforming = true;
-            return (
-              transformer.transformSource(filename, code, false).code || code
-            );
-          } finally {
-            transforming = false;
-          }
-        },
-        {
-          exts: [extname(modulePath)],
-          ignoreNodeModules: false,
-          matcher: (...args) => {
-            if (transforming) {
-              // Don't transform any dependency required by the transformer itself
-              return false;
-            }
-            return transformer.shouldTransform(...args);
-          },
-        },
-      );
-
-      const globalModule = interopRequireDefault(require(modulePath)).default;
+      const globalModule = interopRequireDefault(
+        transformer.requireAndTranspileModule(modulePath),
+      ).default;
 
       if (typeof globalModule !== 'function') {
         throw new TypeError(
@@ -85,8 +56,6 @@ export default async ({
       }
 
       await globalModule(globalConfig);
-
-      revertHook();
     });
   }
